add github workflow to support release and deploy to maven repo

+++ omod/src/main/resources/_etl/derived/vl_episode/facts/sp_mamba_fact_viral_load_episode_create.sql
@@ -55,7 +55,7 @@ CREATE TABLE mamba_fact_viral_load_episode (
     result_numeric_raw DOUBLE,
     result_numeric DOUBLE,
     result_modifier VARCHAR(10),  -- <, >, =
-    result_numeric_copies INT,
+    result_numeric_copies BIGINT,
     lower_detection_limit INT,
 
     -- Qualitative Result
